#2245 Splitting LFRicLoopBounds from dynamo0p3.py
