+++ src/psyclone/domain/lfric/__init__.py
@@ -63,13 +63,16 @@ from psyclone.domain.lfric.metadata_to_arguments_rules import \
 from psyclone.domain.lfric.arg_index_to_metadata_index import \
     ArgIndexToMetadataIndex
 from psyclone.domain.lfric.lfric_collection import LFRicCollection
+from psyclone.domain.lfric.lfric_loop_bounds import LFRicLoopBounds
 
 
 __all__ = [
+    'ArgIndexToMetadataIndex',
     'ArgOrdering',
     'FunctionSpace',
     'KernCallAccArgList',
     'KernCallArgList',
+    'KernCallInvokeArgList',
     'KernelInterface',
     'KernStubArgList',
     'LFRicArgDescriptor',
@@ -77,4 +80,8 @@ __all__ = [
     'LFRicConstants',
     'LFRicExtractDriverCreator',
     'LFRicInvoke',
-    'LFRicSymbolTable']
+    'LFRicKernCallFactory',
+    'LFRicLoopBounds',
+    'LFRicSymbolTable',
+    'LFRicTypes',
+    'MetadataToArgumentsRules']
