+++ src/psyclone/dynamo0p3.py
@@ -3286,69 +3286,6 @@ class DynCellIterators(LFRicCollection):
                 self._first_var.ref_name() + "%get_nlayers()"))
 
 
-class LFRicLoopBounds(LFRicCollection):
-    '''
-    Handles all variables required for specifying loop limits within a
-    PSy-layer routine.
-
-    '''
-    def _invoke_declarations(self, parent):
-        '''
-        Only needed because method is virtual in parent class.
-
-        :param parent: the f2pygen node representing the PSy-layer routine.
-        :type parent: :py:class:`psyclone.f2pygen.SubroutineGen`
-
-        '''
-
-    def initialise(self, parent):
-        '''
-        Updates the f2pygen AST so that all of the variables holding the lower
-        and upper bounds of all loops in an Invoke are initialised.
-
-        :param parent: the f2pygen node representing the PSy-layer routine.
-        :type parent: :py:class:`psyclone.f2pygen.SubroutineGen`
-
-        '''
-        loops = self._invoke.schedule.loops()
-
-        if not loops:
-            return
-
-        parent.add(CommentGen(parent, ""))
-        parent.add(CommentGen(parent, " Set-up all of the loop bounds"))
-        parent.add(CommentGen(parent, ""))
-
-        sym_table = self._invoke.schedule.symbol_table
-        config = Config.get()
-        api_config = config.api_conf("dynamo0.3")
-
-        for idx, loop in enumerate(loops):
-
-            if loop.loop_type == "null":
-                # 'null' loops don't need any bounds.
-                continue
-
-            root_name = f"loop{idx}_start"
-            lbound = sym_table.find_or_create_integer_symbol(root_name,
-                                                             tag=root_name)
-            parent.add(AssignGen(parent, lhs=lbound.name,
-                                 rhs=loop._lower_bound_fortran()))
-            entities = [lbound.name]
-
-            if loop.loop_type != "colour":
-                root_name = f"loop{idx}_stop"
-                ubound = sym_table.find_or_create_integer_symbol(root_name,
-                                                                 tag=root_name)
-                entities.append(ubound.name)
-                parent.add(AssignGen(parent, lhs=ubound.name,
-                                     rhs=loop._upper_bound_fortran()))
-
-            parent.add(DeclGen(parent, datatype="integer",
-                               kind=api_config.default_kind["integer"],
-                               entity_decls=entities))
-
-
 class LFRicScalarArgs(LFRicCollection):
     '''
     Handles the declarations of scalar kernel arguments appearing in either
